Split QuestState to file
