+++ UniPlogger/Model/Quest.swift
@@ -8,24 +8,6 @@
 
 import Foundation
 
-enum QuestState {
-    case todo
-    case doing
-    case done
-    case abandon
-}
-
-extension QuestState: CustomStringConvertible {
-    var description: String {
-        switch self {
-            case .todo: return "To Do"
-            case .doing: return "Doing"
-            case .done: return "Done"
-            case .abandon: return "Abandon"
-        }
-    }
-}
-
 struct Quest {
     var id: String
     var title: String
